Third Umpire Decision Review System (DRS Gully Cricket)
Project is going to be very helpful for those cricket fans who want to create their own decision review system. 
The third empire takes the decision about no-ball, runout, or catch out. Here we are going to make the decision review system using python which will take an accurate decision about whether the batsman is out or not out.